+++ framework/Furion/DatabaseAccessor/ContextPools/DbContextPool.cs
@@ -4,7 +4,7 @@
 //
 // 框架名称：Furion
 // 框架作者：百小僧
-// 框架版本：2.10.8
+// 框架版本：2.11.0
 // 源码地址：Gitee： https://gitee.com/dotnetchina/Furion
 //          Github：https://github.com/monksoul/Furion
 // 开源协议：Apache-2.0（https://gitee.com/dotnetchina/Furion/blob/master/LICENSE）
@@ -12,6 +12,7 @@
 
 using Furion.DependencyInjection;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Storage;
 using StackExchange.Profiling;
 using StackExchange.Profiling.Data;
@@ -111,8 +112,8 @@ namespace Furion.DatabaseAccessor
                         dynamic context = s as DbContext;
 
                         // 当前事务
-                        var database = context.Database;
-                        var currentTransaction = database.CurrentTransaction;
+                        var database = context.Database as DatabaseFacade;
+                        var currentTransaction = database?.CurrentTransaction;
                         if (currentTransaction != null && context.FailedAutoRollback == true)
                         {
                             // 获取数据库连接信息